requirements/airflow.txt

apache-airflow-providers-standard
requests>=2.31,<3
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
PyYAML==6.0.2
matplotlib==3.9.2
mlflow>=2.16,<3
torch==2.4.1
torchvision==0.19.1
dvc==3.57.0
